pyproject.toml
[project]
name = "gamerpssl"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12.6,<4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg (>=3.2.5,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "environ (>=1.0,<2.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "mock (>=5.1.0,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
